Move blocklist to content.
